pyproject.toml
[project]
authors = [
    {name = "Yurasov Aleksey", github = "https://github.com/YurasovAleksey"},
]
name = "hexlet-code"
version = "0.1.0"
description = "Page analyzer project"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "gunicorn>=23.0.0",
    "Flask>=3.1.1",
    "dotenv>=0.9.9",
    "psycopg2-binary>=2.9.9"
]
